[Net] Fix get_response_body_length for large files.
Parsing was fixed, but not the return value for the exposed getter.

(cherry picked from commit 01e5e98312389f09320f72b34e7be4f5f39f19da)

+++ platform/javascript/http_client_javascript.cpp
@@ -185,7 +185,7 @@ Error HTTPClient::get_response_headers(List<String> *r_response) {
 	return OK;
 }
 
-int HTTPClient::get_response_body_length() const {
+int64_t HTTPClient::get_response_body_length() const {
 	return godot_js_fetch_body_length_get(js_id);
 }
